if config_max_items is set to 0 then disable any limits for feed items count, more debug & status messages, implement all_reload function, start fixing statement creation for tags set (this is awful), overall cleanup

+++ src/config.c
@@ -1,8 +1,7 @@
-#include <stdbool.h>
 #include <stddef.h>
-#include <stdint.h>
+#include <stdbool.h>
 
-int64_t config_max_items = 100;
+size_t config_max_items = 100; // 0 == inf
 size_t config_init_parser_buf_size = 100000;
 
 bool config_menu_show_number = true;
